atualizao medico

+++ src/main/java/medi/voli/api/medico/MedicoRepository.java
@@ -2,6 +2,6 @@ package medi.voli.api.medico;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MedicoRepository extends JpaRepository<Medico, Integer>{
+public interface MedicoRepository extends JpaRepository<Medico, Long>{
 
 }
